+++ packages/terra-date-picker/tests/jest/DateInput.test.jsx
@@ -19,7 +19,6 @@ it('should render a default date input with all props', () => {
       onChange={() => {}}
       onClick={() => {}}
       onKeyDown={() => {}}
-      aria-label="date"
     />
   ));